updated openzeplin to older version

+++ contracts/interfaces/IKyberNetworkProxyInterface.sol
@@ -1,22 +1,35 @@
 pragma solidity ^0.5.0;
 
-import "openzeppelin-solidity/contracts/token/ERC20/IERC20.sol";
+import "https://github.com/OpenZeppelin/openzeppelin-contracts/blob/release-v2.4.0/contracts/token/ERC20/IERC20.sol";
 
 interface IKyberNetworkProxyInterface {
-    function maxGasPrice() external view returns(uint);
-    function getUserCapInWei(address user) external view returns(uint);
-    function getUserCapInTokenWei(address user, IERC20 token) external view returns(uint);
-    function enabled() external view returns(bool);
-    function info(bytes32 id) external view returns(uint);
-    function getExpectedRate(IERC20 src, IERC20 dest, uint srcQty)
-        external view returns (uint expectedRate, uint slippageRate);
+    function maxGasPrice() external view returns (uint256);
+
+    function getUserCapInWei(address user) external view returns (uint256);
+
+    function getUserCapInTokenWei(address user, IERC20 token)
+        external
+        view
+        returns (uint256);
+
+    function enabled() external view returns (bool);
+
+    function info(bytes32 id) external view returns (uint256);
+
+    function getExpectedRate(
+        IERC20 src,
+        IERC20 dest,
+        uint256 srcQty
+    ) external view returns (uint256 expectedRate, uint256 slippageRate);
+
     function tradeWithHint(
         IERC20 src,
-        uint srcAmount,
+        uint256 srcAmount,
         IERC20 dest,
         address destAddress,
-        uint maxDestAmount,
-        uint minConversionRate,
+        uint256 maxDestAmount,
+        uint256 minConversionRate,
         address walletId,
-        bytes calldata hint) external payable returns(uint);
+        bytes calldata hint
+    ) external payable returns (uint256);
 }
